pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["Midnull <Midnull@DESKTOP>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
sqladmin = "^0.17.0"
pydantic-settings = "^2.3.3"
uuid6 = "^2024.1.12"
icecream = "^2.1.3"
uvicorn  = {extras = ["all"], version = "^0.30.1"}
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
